pyproject.toml
[project]
name = "clever-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "httpx>=0.28.1",
    "openai>=1.65.1",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ruff>=0.9.9",
]
